Move pylxd.certificate to pylxd.deprecated.certificate
