+++ pylxd/deprecated/certificate.py
